+++ mycodo/scripts/upgrade_post.sh
@@ -26,10 +26,6 @@ TIMER_START_update_swap_size=$SECONDS
 ${INSTALL_CMD} update-swap-size
 TIMER_TOTAL_update_swap_size=$((SECONDS - TIMER_START_update_swap_size))
 
-TIMER_START_setup_virtualenv=$SECONDS
-${INSTALL_CMD} setup-virtualenv
-TIMER_TOTAL_setup_virtualenv=$((SECONDS - TIMER_START_setup_virtualenv))
-
 TIMER_START_update_apt=$SECONDS
 ${INSTALL_CMD} update-apt
 TIMER_TOTAL_update_apt=$((SECONDS - TIMER_START_update_apt))
@@ -53,6 +49,10 @@ TIMER_TOTAL_update_pip3=$((SECONDS - TIMER_START_update_pip3))
 TIMER_START_update_pip3_packages=$SECONDS
 ${INSTALL_CMD} update-pip3-packages
 TIMER_TOTAL_update_pip3_packages=$((SECONDS - TIMER_START_update_pip3_packages))
+
+TIMER_START_setup_virtualenv=$SECONDS
+${INSTALL_CMD} setup-virtualenv
+TIMER_TOTAL_setup_virtualenv=$((SECONDS - TIMER_START_setup_virtualenv))
 
 TIMER_START_update_influxdb=$SECONDS
 ${INSTALL_CMD} update-influxdb