+++ CHANGELOG.md
@@ -105,6 +105,8 @@ Additionally, if you are currently using the MQTT JSON Input and your topics con
  - Place Output columns at back of Graph Widget charts
  - Add Measurements/Units: Specific Gravity, Salinity, Total Dissolved Solids, Parts per Thousand
  - Add conversions for Parts per Thousand
+ - Specify virtualenv install version in requirements.txt
+ - Enable server-side Flask session
 
 
 ## 8.11.0 (2021-06-05)